refactor(v4): move base config to @chakra-ui/panda-preset (Tier A)
Relocate Chakra's design-system definition out of @chakra-ui/react into the
preset package where it belongs: defaultBaseConfig/defaultConditions (base-config),
plus the color-mix and css-var helpers they need. panda-preset's def gains
defineConditions and re-exports the moved pieces.

- @chakra-ui/react imports defaultBaseConfig/colorMix/cssVar from
  @chakra-ui/panda-preset (now a real dependency, not devDep); its ./preset-base
  export is removed.
- @chakra-ui/styled-system's panda.config imports @chakra-ui/panda-preset instead
  of reaching into ../react/src/preset-base — no more cross-package relative hack.
- sandbox config likewise imports @chakra-ui/panda-preset.
- preset.ts bridges the design-system Config to Chakra's SystemConfig at the merge.

Verified: react typechecks at 0; sandbox renders unchanged (style props, Button
recipe, Card slot recipe, patterns all intact).
