+++ packages/react/src/index.ts
@@ -5,5 +5,5 @@ export * from "./hooks"
 export { mergeProps } from "./merge-props"
 export { mergeRefs } from "./merge-refs"
 export { defaultConfig, defaultSystem } from "./preset"
-export { defaultBaseConfig } from "./preset-base"
+export { defaultBaseConfig } from "@chakra-ui/panda-preset"
 export * from "./styled-system"
