+++ packages/react/src/preset.ts
@@ -1,8 +1,11 @@
-import { defaultBaseConfig } from "./preset-base"
-import { createSystem, mergeConfigs } from "./styled-system"
+import { defaultBaseConfig } from "@chakra-ui/panda-preset"
+import { type SystemConfig, createSystem, mergeConfigs } from "./styled-system"
 import { defaultThemeConfig } from "./theme"
 
-export const defaultConfig = mergeConfigs(defaultBaseConfig, defaultThemeConfig)
+export const defaultConfig = mergeConfigs(
+  defaultBaseConfig as SystemConfig,
+  defaultThemeConfig,
+)
 
 export const defaultSystem = createSystem(defaultConfig)
